+++ packages/connector/src/methods.ts
@@ -2,9 +2,11 @@
 import { getUpdateStateDetails, getUpdateStateTx } from './helpers';
 import { MetamaskSnap } from './snap';
 import {
+  CheckCredentialExistenceRequestParams,
   CreateProofRequestParams,
   RPCMethods,
   SaveCredentialsRequestParams,
+  SaveCredentialsResponse,
   W3CCredential,
   ZKPProofResponse,
   ZKPProofSnapResponse,
@@ -23,6 +25,16 @@ const sendSnapMethod = async <T>(
   });
 };
 
+export const checkCredentialExistence = async function (
+  this: MetamaskSnap,
+  params: CheckCredentialExistenceRequestParams,
+): Promise<SaveCredentialsResponse[]> {
+  return await sendSnapMethod(
+    { method: RPCMethods.CheckCredentialExistence, params },
+    this.snapId,
+  );
+};
+
 export const createIdentity = async function (
   this: MetamaskSnap,
 ): Promise<{
@@ -38,7 +50,7 @@ export const createIdentity = async function (
 export const saveCredentials = async function (
   this: MetamaskSnap,
   params: SaveCredentialsRequestParams,
-): Promise<W3CCredential[]> {
+): Promise<SaveCredentialsResponse[]> {
   return await sendSnapMethod(
     { method: RPCMethods.SaveCredentials, params },
     this.snapId,